refactor: Restructure database storage with separate modules for bincode and encrypted bincode stores
